+++ Assets/Prefabs/Player.prefab
@@ -108,6 +108,7 @@ GameObject:
   - component: {fileID: 1088682975866835317}
   - component: {fileID: 3489657001170142623}
   - component: {fileID: 2174348033361710420}
+  - component: {fileID: -4998809826640853646}
   m_Layer: 0
   m_Name: Player
   m_TagString: Untagged
@@ -300,6 +301,18 @@ BoxCollider:
   serializedVersion: 3
   m_Size: {x: 1.36, y: 1, z: 2.6}
   m_Center: {x: 0, y: 0.5, z: 0}
+--- !u!114 &-4998809826640853646
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 1643351420949137732}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: c13debc1718782f448e25fce741408f7, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
 --- !u!1 &6132438955859910525
 GameObject:
   m_ObjectHideFlags: 0
