+++ Assets/Scripts/UI/PlayerUI.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: c13debc1718782f448e25fce741408f7
